switch tests to 'jest'
also clean up how the tests reference the product code
