+++ clients/ts/FunctionalTests/package.json
@@ -4,12 +4,17 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "dependencies": {},
+  "dependencies": {
+    "@aspnet/signalr": "file:../signalr",
+    "@aspnet/signalr-protocol-msgpack": "file:../signalr-protocol-msgpack"
+  },
   "devDependencies": {
     "@types/debug": "0.0.30",
+    "@types/jasmine": "^2.8.6",
     "@types/node": "^9.4.6",
     "debug": "^3.1.0",
     "es6-promise": "^4.2.2",
+    "jasmine": "^3.1.0",
     "tap-parser": "^7.0.0",
     "tee": "^0.2.0",
     "ts-node": "^4.1.0"
