+++ clients/ts/FunctionalTests/ts/TestLogger.ts
@@ -2,7 +2,7 @@ import { ILogger, LogLevel } from "@aspnet/signalr";
 
 // Since JavaScript modules are file-based, we can just pull in utilities from the
 // main library directly even if they aren't exported.
-import { ConsoleLogger } from "../../signalr/src/Utils";
+import { ConsoleLogger } from "@aspnet/signalr/dist/esm/Utils";
 
 export class TestLog {
     public messages: Array<[Date, LogLevel, string]> = [];
