+++ clients/ts/FunctionalTests/tsconfig.json
@@ -1,27 +1,12 @@
 {
-  "compileOnSave": false,
+  "extends": "../tsconfig.base.json",
   "compilerOptions": {
-    "noImplicitAny": false,
-    "noEmitOnError": true,
-    "removeComments": false,
-    "sourceMap": true,
-    "target": "es5",
-    "module": "es2015",
     "outDir": "./obj/js",
-    "baseUrl": ".",
-    "paths": {
-      "@aspnet/signalr": [ "../signalr/dist/esm/index" ],
-      "@aspnet/signalr-protocol-msgpack": [ "../signalr-protocol-msgpack/dist/esm/index" ]
-    },
-    "lib": [ "es2015.promise", "es5", "dom", "es2015.collection" ]
+    "typeRoots": [
+      "./node_modules/@types/"
+    ]
   },
   "include": [
-    "./ts/**/*",
-    "../signalr/dist/esm/**/*.d.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "wwwroot",
-    "js"
+    "./ts/**/*"
   ]
 }
