--- clients/ts/signalr-protocol-msgpack/spec/tsconfig.json
@@ -1,18 +1,0 @@
-{
-    "compileOnSave": false,
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es5",
-        "moduleResolution": "node",
-        "outDir": "./obj",
-        "lib": [ "es2015", "dom" ],
-        "baseUrl": ".",
-        "paths": {
-            "@aspnet/*": [ "../../*" ]
-        }
-    },
-    "include": [
-        "./**/*",
-        "../../typings/**/*"
-    ]
-}
