+++ clients/ts/signalr-protocol-msgpack/tsconfig.json
@@ -1,15 +1,7 @@
 {
-    "extends": "../tsconfig-base.json",
-    "compilerOptions": {
-        "baseUrl": ".",
-        "paths": {
-            "@aspnet/signalr": [
-                "../signalr"
-            ]
-        }
-    },
+    "extends": "../tsconfig.base.json",
     "include": [
-        "./src/**/*",
-        "../typings/**/*"
+        "./src/**/*"
     ]
 }
+
