--- clients/ts/signalr/spec/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compileOnSave": false,
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es2016",
-        "sourceMap": true,
-        "moduleResolution": "node",
-        "outDir": "./obj",
-        "lib": [ "es2016", "dom" ]
-    },
-    "include": [
-        "./**/*",
-        "../../typings/**/*"
-    ]
-}
