+++ clients/ts/signalr/tests/Common.ts
