+++ clients/ts/signalr/tests/HttpClient.test.ts
@@ -3,7 +3,6 @@
 
 import { HttpRequest } from "../src/HttpClient";
 import { TestHttpClient } from "./TestHttpClient";
-import { asyncit as it } from "./Utils";
 
 describe("HttpClient", () => {
     describe("get", () => {
