+++ clients/ts/signalr/tests/TextMessageFormat.test.ts
@@ -17,10 +17,10 @@ describe("TextMessageFormat", () => {
     });
 
     ([
-        ["", new Error("Message is incomplete.")],
-        ["ABC", new Error("Message is incomplete.")],
-        ["ABC\u001eXYZ", new Error("Message is incomplete.")],
-    ] as Array<[string, Error]>).forEach(([payload, expectedError]) => {
+        ["", "Message is incomplete."],
+        ["ABC", "Message is incomplete."],
+        ["ABC\u001eXYZ", "Message is incomplete."],
+    ] as Array<[string, string]>).forEach(([payload, expectedError]) => {
         it(`should fail to parse '${payload}'`, () => {
             expect(() => TextMessageFormat.parse(payload)).toThrow(expectedError);
         });
