+++ clients/ts/signalr/tests/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "include": [
+        "./**/*"
+    ]
+}
