+++ clients/ts/signalr/tsconfig.json
@@ -1,5 +1,5 @@
 {
-    "extends": "../tsconfig-base.json",
+    "extends": "../tsconfig.base.json",
     "include": [
         "./src/**/*"
     ]
