+++ clients/ts/tsconfig.base.json
@@ -2,10 +2,10 @@
     "compileOnSave": false,
     "compilerOptions": {
         "module": "es2015",
-        "target": "es2016",
-        "outDir": "./obj/js",
+        "target": "es5",
         "sourceMap": true,
         "moduleResolution": "node",
+        "importHelpers": true,
         "inlineSources": true,
         "emitDecoratorMetadata": true,
         "experimentalDecorators": true,
@@ -14,6 +14,15 @@
         "suppressImplicitAnyIndexErrors": true,
         "noEmitOnError": true,
         "stripInternal": true,
-        "lib": [ "es5", "es2015.promise", "es2015.iterable", "dom" ]
+        "lib": [ "es5", "es2015.promise", "es2015.iterable", "dom" ],
+        "baseUrl": ".",
+        "paths": {
+            "@aspnet/signalr": [
+                "./signalr"
+            ],
+            "@aspnet/signalr-protocol-msgpack": [
+                "./signalr-protocol-msgpack"
+            ]
+        }
     }
 }
