+++ clients/ts/tsconfig.jest.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./tsconfig.base.json",
+    "compilerOptions": {
+        "module": "commonjs"
+    },
+    "include": [
+        "./*/tests/**/*"
+    ]
+}
